+++ nn/runtime/cpu/q16/cpu_q16_softmax.c
@@ -31,7 +31,7 @@ int layer_cpu_q16_SOFTMAX_init(const nn_t* nn, const layer_t* layer)
 	}
 	else
 	{
-		r = rte_cpu_create_layer_common(nn, layer, sizeof(layer_cpu_q16_softmax_context_t), sizeof(int8_t));
+		r = rte_cpu_create_layer_common(nn, layer, sizeof(layer_cpu_q16_softmax_context_t), sizeof(int16_t));
 	}
 
 	if(0 == r)
